Refactor packages, names, and add logging mechanisims
simplified the package names and structure
re-wired all the tests
added in the logging mechanisms but it's not logging
yet because nothing is memoized
