+++ snappy/common/fixtures.py
