--- snappy/load_config.py
@@ -1,6 +1,0 @@
-from swift_silent_deadly.swift_config import ObjectStorageAPIConfig
-
-
-config = ObjectStorageAPIConfig(
-    config_file_path='dfw.config',
-    section_name=ObjectStorageAPIConfig.SECTION_NAME)
