+++ snappy/swift_config.py
