+++ snappy/swift_constants.py
