+++ snappy/tests/features/cors_test.py
@@ -2,9 +2,9 @@ from snappy.common.decorators import (DataDrivenFixture, data_driven_test)
 
 from behest.client import HTTPClient
 
-from snappy.constants import Constants
-from snappy.fixtures.swift_fixtures import ObjectStorageFixture
-from snappy.generators import ObjectDatasetList
+from snappy.swift_constants import Constants
+from snappy.swift_fixtures import ObjectStorageFixture
+from snappy.swift_generators import ObjectDatasetList
 
 CONTAINER_DESCRIPTOR = 'cors_container'
