+++ snappy/tests/regression/container/obj_list_extended_accept_header_test.py
@@ -1,6 +1,6 @@
 from snappy.tools import randomstring as randstring
 from snappy.common.decorators import (DataDrivenFixture, data_driven_test)
-from snappy.fixtures.swift_fixtures import ObjectStorageFixture
+from snappy.swift_fixtures import ObjectStorageFixture
 from snappy.common.datasets import DatasetList
 
 CONTENT_TYPE_TEXT = 'text/plain; charset=UTF-8'
