+++ snappy/tests/regression/object/object_versioning.py
@@ -1,6 +1,6 @@
 from snappy.tools import randomstring as randstring
-from snappy.constants import Constants
-from snappy.fixtures.swift_fixtures import ObjectStorageFixture
+from snappy.swift_constants import Constants
+from snappy.swift_fixtures import ObjectStorageFixture
 
 CONTENT_TYPE_TEXT = "text/plain; charset=UTF-8"
 CONTAINER_NAME = "object_versioning_test"
