+++ snappy/tests/smoke/quick_test.py
@@ -1,9 +1,9 @@
 from snappy.common.decorators import (DataDrivenFixture, data_driven_test)
-from snappy.generators import ObjectDatasetList
-from snappy.constants import Constants
+from snappy.swift_generators import ObjectDatasetList
+from snappy.swift_constants import Constants
 
 
-from snappy.fixtures.swift_fixtures import ObjectStorageFixture
+from snappy.swift_fixtures import ObjectStorageFixture
 
 @DataDrivenFixture
 class QuickTest(ObjectStorageFixture):
